Possible fix to XP per kill rounding/truncation issue
The multiplication step '(ownerlevel * 5 + nBaseExp) * BaseGainLevelFactor()' produces a 'float' value, but since the return type of 'BaseGain' is 'uint32', the compiler implicitly truncates the XP value when converting from 'float' to 'uint32'.

By changing the return type of 'BaseGain' from 'uint32' to 'float', this preserves the fractional XP value until the very end, when the XP value is converted to an integer using 'std:nearbyint()'.

I did very limited testing, but killing a Level 3 mob as Level 2 player did, in fact, yield '58' rather than '57' XP.

+++ src/game/Tools/Formulas.h
@@ -129,7 +129,7 @@ namespace MaNGOS
             return 17;
         }
 
-        inline uint32 BaseGain(uint32 unit_level, uint32 mob_level, ContentLevels content)
+        inline float BaseGain(uint32 unit_level, uint32 mob_level, ContentLevels content)
         {
             uint32 nBaseExp = unit_level * 5;
             switch (content)
